fix: - DB 재설정 - 파일 구조 변경 src 삭제 및 nodemon path 변경
