--- src/database/config.js
@@ -1,11 +1,0 @@
-import dotenv from "dotenv";
-dotenv.config();
-
-const config = {
-  MYSQL_HOST: process.env.MYSQL_HOST,
-  MYSQL_PORT: Number(process.env.MYSQL_PORT),
-  MYSQL_USER: process.env.MYSQL_USER,
-  MYSQL_PASSWORD: process.env.MYSQL_PASSWORD,
-  MYSQL_DB: process.env.MYSQL_DB,
-};
-export default config;
