--- src/middlewares/authMiddleware.js
@@ -1,17 +1,0 @@
-import jwt from 'jsonwebtoken';
-
-export const verifyJWT = (req, res, next) => {
-  const token = req.headers['authorization']?.split(' ')[1];
-
-  if (!token) {
-    return res.status(403).json({ message: 'A token is required for authentication' });
-  }
-
-  jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
-    if (err) {
-      return res.status(401).json({ message: 'Invalid Token' });
-    }
-    req.user = decoded;
-    next();
-  });
-};
